refactor: ElasticSearch component
