+++ src/pages/BankCard/BankCard.tsx
@@ -1,6 +1,6 @@
 import { PureComponent } from 'react';
 
-import ElasticSearch from '@/components/ElasticSearch';
+import { ElasticSearch } from '@/components/ElasticSearch';
 import MyMap from '@/components/MyMap';
 import UpdateDate from '@/components/UpdateDate';
 import { BANKS_DATA } from '@/constants';
